Kenny2

+++ app/src/main/java/com/example/runqr/Location.java
@@ -27,7 +27,7 @@ public class Location{
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
@@ -36,7 +36,7 @@ public class Location{
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
